requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.1
PyJWT==2.8.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.0
```

**Procfile:**
```
web: gunicorn app:app
```

**runtime.txt:**
```
python-3.11.7